divided Services.test.js to three separate test files for each model class
